Extracting general Retrofit factory to be utilized in other services
